commid 13-05-2020 login

+++ database/migrations/2020_04_25_070053_crear_tabla_menu_rol.php
@@ -17,7 +17,7 @@ class CrearTablaMenuRol extends Migration
             $table->unsignedBigInteger('rol_id');
             $table->foreign('rol_id','fk_menurol_rol')->references('id')->on('rol')
             ->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedInteger('menu_id');
+            $table->unsignedBigInteger('menu_id');
             $table->foreign('menu_id','fk_menurol_menu')->references('id')->on('menu')
             ->onDelete('restrict')->onUpdate('restrict');
             $table->charset='utf8mb4';
